Move common model utils functions to models_utils.py
