+++ forge/test/models/onnx/multimodal/oft/test_oft_onnx.py
@@ -6,7 +6,8 @@ import pytest
 import forge
 from forge.verify.verify import verify
 
-from test.models.onnx.multimodal.oft.utils.oft_utils import get_inputs, get_models
+from test.models.pytorch.multimodal.oft.utils.oft_utils import get_inputs
+from test.models.onnx.multimodal.oft.utils.oft_utils import get_models
 from forge.forge_property_utils import Framework, Source, Task
