+++ forge/test/models/onnx/text/bi_lstm_crf/test_bilstm_crf.py
@@ -9,7 +9,7 @@ import onnx
 import forge
 from forge.verify.verify import verify
 
-from test.models.onnx.text.bi_lstm_crf.utils.model import get_model
+from test.models.pytorch.text.bi_lstm_crf.utils.model import get_model
 from forge.forge_property_utils import Framework, Source, Task
