--- forge/test/models/onnx/vision/mobilenetv2/utils/__init__.py
@@ -1,4 +1,0 @@
-# SPDX-FileCopyrightText: (c) 2025 Tenstorrent AI ULC
-#
-# SPDX-License-Identifier: Apache-2.0
-from .utils import load_inputs
