--- forge/test/models/onnx/vision/utils/__init__.py
@@ -1,4 +1,0 @@
-# SPDX-FileCopyrightText: (c) 2024 Tenstorrent AI ULC
-#
-# SPDX-License-Identifier: Apache-2.0
-from .utils import load_inputs
