+++ forge/test/models/pytorch/text/gemma/test_gemma_v1.py
@@ -8,10 +8,7 @@ import forge
 from forge.forge_property_utils import Framework, Source, Task
 from forge.verify.verify import verify
 
-from test.models.pytorch.text.gemma.utils.model_utils import (
-    generate_no_cache,
-    pad_inputs,
-)
+from test.models.models_utils import generate_no_cache, pad_inputs
 
 
 @pytest.mark.nightly
