--- forge/test/models/pytorch/text/gemma/utils/model_utils.py
@@ -1,52 +1,0 @@
-# SPDX-FileCopyrightText: (c) 2025 Tenstorrent AI ULC
-#
-# SPDX-License-Identifier: Apache-2.0
-import torch
-
-
-def generate_no_cache(max_new_tokens, model, inputs, seq_len, tokenizer):
-    """
-    Generates text autoregressively without using a KV cache, iteratively predicting one token at a time.
-    The function stops generation if the maximum number of new tokens is reached or an end-of-sequence (EOS) token is encountered.
-
-    Args:
-        max_new_tokens (int): The maximum number of new tokens to generate.
-        model (torch.nn.Module): The language model used for token generation.
-        inputs (torch.Tensor): Input tensor of shape (batch_size, seq_len), representing tokenized text.
-        seq_len (int): The current sequence length before generation starts.
-        tokenizer: The tokenizer used to decode token IDs into text.
-
-    Returns:
-        str: The generated text after decoding the new tokens.
-    """
-    current_pos = seq_len
-
-    for _ in range(max_new_tokens):
-        logits = model(inputs)
-        # Get only the logits corresponding to the last valid token
-        if isinstance(logits, (list, tuple)):
-            logits = logits[0]
-
-        next_token_logits = logits[:, current_pos - 1, :]
-        next_token_id = torch.argmax(next_token_logits, dim=-1)
-        # Stop if EOS token is encountered
-        if next_token_id.item() == tokenizer.eos_token_id:
-            break
-
-        inputs[:, current_pos] = next_token_id
-
-        current_pos += 1  # Move to next position
-
-    # Decode valid tokens
-    valid_tokens = inputs[:, seq_len:current_pos].view(-1).tolist()
-    answer = tokenizer.decode(valid_tokens, skip_special_tokens=True)
-
-    return answer
-
-
-def pad_inputs(inputs, max_new_tokens):
-    batch_size, seq_len = inputs.shape
-    max_seq_len = seq_len + max_new_tokens
-    padded_inputs = torch.zeros((batch_size, max_seq_len), dtype=inputs.dtype, device=inputs.device)
-    padded_inputs[:, :seq_len] = inputs
-    return padded_inputs, seq_len
