+++ forge/test/models/pytorch/vision/mgp_str_base/test_mgp_str_base.py
@@ -10,7 +10,7 @@ import forge
 from forge.forge_property_utils import Framework, Source, Task
 from forge.verify.verify import DepricatedVerifyConfig, verify
 
-from test.models.pytorch.vision.mgp_str_base.utils.utils import load_input, load_model
+from test.models.models_utils import load_input, load_model
 
 
 class Wrapper(torch.nn.Module):
