+++ forge/test/models/pytorch/vision/mobilenet/test_mobilenet_v1.py
@@ -2,20 +2,18 @@
 
 # SPDX-License-Identifier: Apache-2.0
 import pytest
-import requests
-from PIL import Image
-from transformers import AutoImageProcessor, AutoModelForImageClassification
 
 import forge
 from forge.forge_property_utils import Framework, Source, Task
 from forge.verify.verify import verify
 
-from test.models.pytorch.vision.mobilenet.utils.utils import (
-    load_mobilenet_model,
+from test.models.models_utils import (
+    load_input,
+    load_model,
+    load_timm_model_and_input,
     post_processing,
 )
-from test.models.pytorch.vision.utils.utils import load_timm_model_and_input
-from test.utils import download_model
+from test.models.pytorch.vision.mobilenet.utils.utils import load_mobilenet_model
 
 
 @pytest.mark.nightly
@@ -48,22 +46,6 @@ def test_mobilenetv1_basic(forge_property_recorder):
     post_processing(co_out)
 
 
-def generate_model_mobilenetv1_imgcls_hf_pytorch(variant):
-    # Create Forge module from PyTorch model
-    preprocessor = download_model(AutoImageProcessor.from_pretrained, variant)
-    model = download_model(AutoModelForImageClassification.from_pretrained, variant)
-    # tt_model = forge.PyTorchModule("mobilenet_v1__hf_075_192", model)
-
-    # Image load and pre-processing into pixel_values
-    url = "http://images.cocodataset.org/val2017/000000039769.jpg"
-    image = Image.open(requests.get(url, stream=True).raw)
-    inputs = preprocessor(images=image, return_tensors="pt")
-
-    image_tensor = inputs.pixel_values
-
-    return model, [image_tensor], {}
-
-
 @pytest.mark.nightly
 @pytest.mark.parametrize("variant", ["google/mobilenet_v1_0.75_192"])
 def test_mobilenetv1_192(forge_property_recorder, variant):
@@ -81,7 +63,9 @@ def test_mobilenetv1_192(forge_property_recorder, variant):
     # Record Forge Property
     forge_property_recorder.record_group("generality")
 
-    framework_model, inputs, _ = generate_model_mobilenetv1_imgcls_hf_pytorch(variant)
+    # Load model and input
+    framework_model = load_model(variant)
+    inputs = load_input(variant)
 
     # Forge compile framework model
     compiled_model = forge.compile(
@@ -90,21 +74,6 @@ def test_mobilenetv1_192(forge_property_recorder, variant):
 
     # Model Verification
     verify(inputs, framework_model, compiled_model, forge_property_handler=forge_property_recorder)
-
-
-def generate_model_mobilenetV1I224_imgcls_hf_pytorch(variant):
-    # Create Forge module from PyTorch model
-    preprocessor = download_model(AutoImageProcessor.from_pretrained, variant)
-    model = download_model(AutoModelForImageClassification.from_pretrained, variant)
-
-    # Image load and pre-processing into pixel_values
-    url = "http://images.cocodataset.org/val2017/000000039769.jpg"
-    image = Image.open(requests.get(url, stream=True).raw)
-    inputs = preprocessor(images=image, return_tensors="pt")
-
-    image_tensor = inputs.pixel_values
-
-    return model, [image_tensor], {}
 
 
 @pytest.mark.nightly
@@ -124,7 +93,9 @@ def test_mobilenetv1_224(forge_property_recorder, variant):
     # Record Forge Property
     forge_property_recorder.record_group("generality")
 
-    framework_model, inputs, _ = generate_model_mobilenetV1I224_imgcls_hf_pytorch(variant)
+    # Load model and input
+    framework_model = load_model(variant)
+    inputs = load_input(variant)
 
     # Forge compile framework model
     compiled_model = forge.compile(
