+++ forge/test/models/pytorch/vision/mobilenet/utils/utils.py
@@ -1,7 +1,6 @@
 # SPDX-FileCopyrightText: (c) 2025 Tenstorrent AI ULC
 #
 # SPDX-License-Identifier: Apache-2.0
-import os
 import urllib
 
 import torch
@@ -40,22 +39,3 @@ def load_mobilenet_model(model_name):
     input_batch = input_tensor.unsqueeze(0)
 
     return model, [input_batch]
-
-
-url = "https://raw.githubusercontent.com/pytorch/hub/master/imagenet_classes.txt"
-
-
-def post_processing(output, top_k=5):
-
-    probabilities = torch.nn.functional.softmax(output[0][0], dim=0)
-    urllib.request.urlretrieve(url, "imagenet_classes.txt")
-
-    with open("imagenet_classes.txt", "r") as f:
-        categories = [s.strip() for s in f.readlines()]
-    topk_prob, topk_catid = torch.topk(probabilities, top_k)
-    for i in range(topk_prob.size(0)):
-        print(categories[topk_catid[i]], topk_prob[i].item())
-
-    # Cleanup
-    os.remove("imagenet_classes.txt")
-    os.remove("dog.jpg")
