+++ forge/test/models/pytorch/vision/resnet/test_resnet.py
@@ -16,7 +16,7 @@ from forge.verify.config import VerifyConfig
 from forge.verify.value_checkers import AutomaticValueChecker
 from forge.verify.verify import verify
 
-from test.models.pytorch.vision.utils.utils import load_vision_model_and_input
+from test.models.models_utils import load_vision_model_and_input
 from test.utils import download_model
 
 variants = [
