+++ forge/test/models/pytorch/vision/retinanet/test_retinanet.py
@@ -12,9 +12,9 @@ import forge
 from forge.forge_property_utils import Framework, Source, Task
 from forge.verify.verify import verify
 
+from test.models.models_utils import load_vision_model_and_input
 from test.models.pytorch.vision.retinanet.utils.image_utils import img_preprocess
 from test.models.pytorch.vision.retinanet.utils.model import Model
-from test.models.pytorch.vision.utils.utils import load_vision_model_and_input
 
 variants = [
     "retinanet_rn18fpn",
