--- forge/test/models/pytorch/vision/segformer/utils/image_utils.py
@@ -1,15 +1,0 @@
-# SPDX-FileCopyrightText: © 2024 Tenstorrent AI ULC
-
-# SPDX-License-Identifier: Apache-2.0
-import requests
-from PIL import Image
-from transformers import AutoImageProcessor
-
-
-def get_sample_data(model_name):
-    url = "http://images.cocodataset.org/val2017/000000039769.jpg"
-    image = Image.open(requests.get(url, stream=True).raw)
-
-    image_processor = AutoImageProcessor.from_pretrained(model_name)
-    pixel_values = image_processor(images=image, return_tensors="pt").pixel_values
-    return pixel_values
