+++ forge/test/models/pytorch/vision/ssd300_vgg16/test_ssd300_vgg16.py
@@ -7,7 +7,7 @@ import forge
 from forge.forge_property_utils import Framework, Source, Task
 from forge.verify.verify import verify
 
-from test.models.pytorch.vision.utils.utils import load_vision_model_and_input
+from test.models.models_utils import load_vision_model_and_input
 
 variants_with_weights = {
     "ssd300_vgg16": "SSD300_VGG16_Weights",
