+++ forge/test/models/pytorch/vision/swin/test_swin.py
@@ -15,8 +15,8 @@ import forge
 from forge.forge_property_utils import Framework, Source, Task
 from forge.verify.verify import verify
 
+from test.models.models_utils import load_vision_model_and_input
 from test.models.pytorch.vision.swin.utils.image_utils import load_image
-from test.models.pytorch.vision.utils.utils import load_vision_model_and_input
 
 
 @pytest.mark.nightly
